[quant][graphmode][fx] Optimize cat
Summary:
Previously we have a cat that takes a list of Tensors with different qparams and dequantize them
cacatenate them and requantize with the output qparams. This adds some unnecessary overhead in dequantizing
and quantizing Tensors.

This PR adds an optimization for cat operator, we'll make sure inputs and output of cat
uses same observer/fake_quant and produce a cat that does not do rescaling.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9ce3257df0bbef1cabb86b375f0d36ec78d4bc05
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54813